[FIX] base_search_custom_field_filter: Avoid hiding model fields in other parts of the UI
Before these changes, if a model field was added to the custom filters, it would become disabled in other parts of the interface — for example, when trying to group by that field.

After these changes, the filter for the simple field is added while preserving the ability to group by it. Only fields that don't belong to the data model itself — that is, those using '.' notation — are hidden.